ready to try new game engine

+++ src_old_2/1_view/0_fps/fps.h
@@ -18,7 +18,7 @@ typedef struct s_time
 size_t	fps_ms_to_frames(size_t ms);
 size_t	fps_frames_to_ms(size_t frames);
 bool	fps_should_render(void);
-size_t	fps_time_in_ms(void);
+int		fps_time_in_ms(void);
 void	fps_update_time(t_time *time);
 
 # endif
